Created PlanetsTable component
