+++ src/components/PlanetsTable/PlanetsTable.scss
@@ -1,0 +1,14 @@
+.planets-table {
+  border-collapse: collapse;
+}
+
+.planets-table__row th {
+  padding: 4px 12px;
+  border-top: 2px solid black;
+  border-bottom: 2px solid black;
+}
+
+.planets-table__row td {
+  padding: 4px 8px;
+  border-bottom: 1px solid grey;
+}
