+++ src/index.scss
@@ -25,25 +25,6 @@ h1 {
   align-items: center;
 }
 
-.planets-section__table {
-  margin-bottom: 20px;
-}
-
-.planets-table {
-  border-collapse: collapse;
-}
-
-.planets-table__row th {
-  padding: 4px 12px;
-  border-top: 2px solid black;
-  border-bottom: 2px solid black;
-}
-
-.planets-table__row td {
-  padding: 4px 8px;
-  border-bottom: 1px solid grey;
-}
-
 .pagination {
   width: 100%;
   display: flex;
